fixed constraint indexing bug
