+++ src/js/ui/NewDataExploration/Helpers.js
@@ -4,3 +4,7 @@ import Util from "../../library/apertureUtil";
 export function prettifyJSON(name) {
     return Util.capitalizeString(Util.underScoreToSpace(name));
 }
+
+export function randomizeIndex() {
+    return (Math.random() * 10000);
+}
