+++ src/js/ui/NewDataExploration/WorkspaceLayers.js
@@ -2,6 +2,7 @@ import React from 'react';
 import { makeStyles } from '@material-ui/core/styles';
 import Layer from "./Layer";
 import {isComponentRerendering} from "./Workspace";
+import {randomizeIndex} from "./Helpers";
 
 const useStyles = makeStyles((theme) => ({
     root: {
@@ -9,13 +10,15 @@ const useStyles = makeStyles((theme) => ({
     },
 }));
 
-function createWorkspace(layers, graphableLayers, openLayers, setOpenLayers, workspace, setWorkspace,  layerTitles) {
+function createWorkspace(layers, graphableLayers, openLayers, setOpenLayers, workspace, setWorkspace, layerTitles) {
     let workspaceLayers = [];
     workspace.forEach((layer, index) => {
         if(layer) {
+            const originalIndex = index;
+            index = randomizeIndex();
             workspaceLayers.push(
-                <div id={`layer-div-${index}`}>
-                    <Layer layer={layers[index]} layerIndex={index} graphableLayers={graphableLayers} layerTitles={layerTitles}
+                <div key={index} id={`layer-div-${originalIndex}`}>
+                    <Layer layer={layers[originalIndex]} layerIndex={originalIndex} graphableLayers={graphableLayers} layerTitles={layerTitles}
                            openLayers={openLayers} setOpenLayers={setOpenLayers}
                            workspace={workspace} setWorkspace={setWorkspace} />
                 </div>
@@ -33,8 +36,10 @@ export default function WorkspaceLayers(props) {
     if(isComponentRerendering) {console.log("|WorkspaceLayers Rerending|")}
     return (
         <div className={classes.root}>
-            {workspaceLayers.map((layer, index) =>
-                <div key={index}>{layer}</div>
+            {workspaceLayers.map((layer, index) => {
+                    index = randomizeIndex();
+                    return (<div key={index}>{layer}</div>)
+                }
             )}
         </div>
     );
